requirements.txt
streamlit>=1.48.1
requests>=2.32.5
beautifulsoup4>=4.13.5
trafilatura>=2.0.0
psycopg2-binary>=2.9.10
